pyproject.toml
[tool.poetry]
name = "zippy"
version = "0.1.0"
description = "Automatically unzip and sort zip files from a watch folder"
authors = ["Ethan Mann <emannguitar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
